backend/requirements.txt
flask
flask_sqlalchemy
pymysql
python-dotenv
flask_jwt_extended
werkzeug.security